Move loader classes to their own module, per @jan-janssen's suggestion
